+++ rustyhack_lib/src/background_map.rs
@@ -17,8 +17,8 @@ impl BackgroundMap {
     }
 
     #[must_use]
-    pub fn get_tile_at(&self, x: usize, y: usize) -> Tile {
-        self.data[y][x]
+    pub fn get_tile_at(&self, x: u32, y: u32) -> Tile {
+        self.data[y as usize][x as usize]
     }
 }
