+++ rustyhack_lib/src/background_map/tiles/wall.rs
@@ -10,7 +10,7 @@ pub struct Wall {
 
 impl Wall {
     #[must_use]
-    pub fn new(x: usize, y: usize, character_icon: char) -> Wall {
+    pub fn new(x: u32, y: u32, character_icon: char) -> Wall {
         Wall {
             tile_position: TilePosition { x, y },
             character_icon,
